Extract common code from exporters
